+++ src/commonMain/kotlin/io/github/optimumcode/json/schema/JsonSchemaLoader.kt
@@ -4,7 +4,6 @@ import io.github.optimumcode.json.schema.SchemaType.DRAFT_2019_09
 import io.github.optimumcode.json.schema.SchemaType.DRAFT_2020_12
 import io.github.optimumcode.json.schema.SchemaType.DRAFT_7
 import io.github.optimumcode.json.schema.extension.ExternalAssertionFactory
-import io.github.optimumcode.json.schema.internal.FormatValidator
 import io.github.optimumcode.json.schema.internal.SchemaLoader
 import io.github.optimumcode.json.schema.internal.wellknown.Draft201909
 import io.github.optimumcode.json.schema.internal.wellknown.Draft202012
@@ -54,11 +53,13 @@ public interface JsonSchemaLoader {
 
   public fun withExtensions(externalFactories: Iterable<ExternalAssertionFactory>): JsonSchemaLoader
 
+  @ExperimentalApi
   public fun withCustomFormat(
     format: String,
     formatValidator: FormatValidator,
   ): JsonSchemaLoader
 
+  @ExperimentalApi
   public fun withCustomFormats(formats: Map<String, FormatValidator>): JsonSchemaLoader
 
   public fun fromDefinition(schema: String): JsonSchema = fromDefinition(schema, null)
