+++ src/commonMain/kotlin/io/github/optimumcode/json/schema/internal/formats/DateTimeFormatValidator.kt
@@ -1,7 +1,7 @@
 package io.github.optimumcode.json.schema.internal.formats
 
-import io.github.optimumcode.json.schema.internal.FormatValidationResult
-import io.github.optimumcode.json.schema.internal.FormatValidator
+import io.github.optimumcode.json.schema.FormatValidationResult
+import io.github.optimumcode.json.schema.FormatValidator
 
 internal object DateTimeFormatValidator : AbstractStringFormatValidator() {
   private const val DATE_TIME_SEPARATOR = 'T'
